Code
Currently this model uses black lines to provide itself direction. GPS module is not present currently in the model but there is a possibility of its addition in future. It avoids the black line(using an IR sensor) when detects a pothole using ultrasonic sensors and uses the code to automatically direct it towards the black line after detecting the pothole.